pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "counterfactual-world-engine"
version = "0.1.0"
description = "Agentic system for reconstructing and simulating alternate realities from incident artifacts using VLMs"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Multiverse Team" }
]
keywords = [
    "vlm",
    "multimodal",
    "counterfactual",
    "simulation",
    "causal-inference",
    "agents"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # VLM Providers
    "google-genai>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.40.0",
    
    # API Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=13.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # Data Processing
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "networkx>=3.0",  # Graph operations
    
    # Video Processing
    "opencv-python>=4.9.0",
    "pillow>=10.0.0",
    
    # Document Processing
    "pypdf>=4.0.0",
    "python-docx>=1.0.0",
    
    # Storage & Caching
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.29.0",  # PostgreSQL async
    "redis>=5.0.0",
    "boto3>=1.34.0",  # S3
    
    # Task Queue
    "celery>=5.4.0",
    
    # CLI
    "typer>=0.12.0",
    "rich>=13.0.0",
    
    # Utilities
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
    "tenacity>=8.0.0",  # Retries
    "orjson>=3.10.0",  # Fast JSON
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "hypothesis>=6.100.0",
    "mypy>=1.11.0",
    "ruff>=0.6.0",
    "pre-commit>=3.8.0",
    "ipython>=8.0.0",
]

viz = [
    "plotly>=5.24.0",
    "matplotlib>=3.9.0",
]

vector = [
    "weaviate-client>=4.0.0",
]

all = [
    "counterfactual-world-engine[dev,viz,vector]",
]

[project.scripts]
cwe = "cwe.cli.main:app"

[project.urls]
Homepage = "https://github.com/multiverse/counterfactual-world-engine"
Documentation = "https://github.com/multiverse/counterfactual-world-engine/docs"
Repository = "https://github.com/multiverse/counterfactual-world-engine"

[tool.hatch.build.targets.wheel]
packages = ["cwe"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["cwe"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "cv2.*",
    "celery.*",
    "redis.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=cwe --cov-report=term-missing"

[tool.coverage.run]
source = ["cwe"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
